requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
bootstrap4-cdn==4.0.1
bootstrap4-nav==1.0.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==35.0.0
Django==3.2.8
django-bootstrap4==22.2
docopt==0.6.2
dominate==2.7.0
flash==1.0.3
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-CKEditor==0.4.6
Flask-MySQLdb==1.0.1
google-auth==2.11.0
google-auth-oauthlib==0.5.3
gspread==5.5.0
httplib2==0.20.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauth2client==4.1.3
oauthlib==3.2.1
pipreqs==0.4.11
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.11.0
pyparsing==3.0.9
pytz==2022.4
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
yarg==0.1.9